+++ lib/AST/CMakeLists.txt
@@ -71,12 +71,12 @@ add_swift_host_library(swiftAST STATIC
   PlatformKind.cpp
   PrettyStackTrace.cpp
   ProtocolConformance.cpp
-  ProtocolGraph.cpp
   RawComment.cpp
   RequirementEnvironment.cpp
-  RequirementMachine.cpp
-  RewriteSystem.cpp
-  RewriteSystemCompletion.cpp
+  RequirementMachine/ProtocolGraph.cpp
+  RequirementMachine/RequirementMachine.cpp
+  RequirementMachine/RewriteSystem.cpp
+  RequirementMachine/RewriteSystemCompletion.cpp
   SILLayout.cpp
   Stmt.cpp
   SubstitutionMap.cpp
