+++ lib/AST/RequirementMachine/RewriteSystem.cpp
@@ -10,11 +10,12 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "swift/AST/RewriteSystem.h"
 #include "llvm/ADT/FoldingSet.h"
 #include "llvm/Support/raw_ostream.h"
 #include <algorithm>
 #include <vector>
+
+#include "RewriteSystem.h"
 
 using namespace swift;
 using namespace rewriting;
